feat: added learning plans and enhanced prompts
Added the ability for users to create lesson plan. The lesson plan name
and description are then passed to an LLM which will search the internet
looking for articles and videos to help the user learn what they are
trying to learn.

The users provided title and description can be enhanced by an LLM to
hopefully provide a better lesson plan when passing the name and
description to the larger llm.